Fixed processing of statename-value method of input. For some reason the code tried to pull the index using the experimental state to simulation state link. It has now been fixed to use the list of simulation state names instead.
The processing of the input shouldn't require the experimental state because experimental states are almost always a one to many relationship. Having a single input profile spread over multiple states would be impossible to fully define, which means the model will not be full defined and hence useless. So restricting the input to only a single state makes sense.