Changing the architecture of the program such that the dynamic equation is now lumped inside the reaction rules. The other supporting scripts have also been changed to reflect this.
The net effect of these changes are, the supporting scripts are now truly independent of the dynamic equation, network architecture and parameter definitions. The definition files have been distilled down to only define what a reaction is, how to convert it into a tensor and then how those tensors fit into the dynamic equation. The in between steps are all done by the supporting scripts.