Changed compDis to account for different in compartment size between complex and enzyme/substrate.
Modification of Michaelis Constant due to compartmentalisation no longer hard coded. An additional tensor index has been assigned to treat this.

Compartment value is now implemented as a true variable and no longer needs to be preassigned during the model parsing stage.

Generation of tensor when free parameters are passed as inputs no longer overrides the tensor structure. A separate temporory tensor is created which is used for simulation (last update had this item but the implementation was bugged).

Various updates for the validation cases.