added sigRxnList as a class for the list of reactions. Updated parseModelm and odeKinetic to address this change.
sigRxnList will be designed as the master list for reaction definition in the sigMat program. All reaction rules must be created in accordance to this format.

The field "label" in rxn has been changed to "desc"

Test cases have been been modified to conform to the change from label to desc.