Moved compDis from objFun to odeQSSA folder.
Removed mkRules folder (it is extremely outdated. Probably better to simply rewrite the code).

Removed objFun folder (the script within requires the data structure to be of a very specific format. This is highly inflexible. It is more likely the user will process each piece of data manually, or design a data processing script designed for the type of data they generate. Thus it is not useful to have my own tool.

Fixed model templates for the new version of the code.

Parameter number and vector orientation checkers included in insParam to provide extra feedback to users.