Fixed modelTemplate (Bnd.Comp is now capitalised).
Fixed MCMC such that when the objective function returns NaN, it is not automatically selected (when min([1 NaN]) is performed, it selects 1, which means it gets chosen, so NaN always get selected. This probably makes the program get stuck in NaN areas which doesn't get saved, making the program get stuck.

MCMCSeriel: Changed disp to dispOpts since disp is a real function.

parseModel: Removed insertion of parameter set into model from its functionality because it is now in insParam.