Fixed some commenting issues with odeQSSA.m
Added some code in parseModelm that enables custom parameters in concentrations, and defaults parameters in compartment size.

Added "amount" and "concentration" switches in model definition. So it enables defining proteins as concentrations or amounts (globally though). This means you can say "I know the amount of protein I have is X, but I don't know that size. So the concentration will vary with the compartment size, but amount is some value". The usage is a model file is defined in the beginning with either spcMode = "a" or "c" for the two mentioned cases. This required changes in insParam (actually implement the change) and parseModelm (include as model definition).

Gave the "model parsed is already a struct" warning in parseModel

Switched off "model entered is already passed" warning in MCMC, so it doesn't keep throwing errors.