Added sorting function to combineResults.m
Attempted to fix another parallel computing freezing bug. This time no slaves terminated at all. The suspicion is the master-lab enters the final receive steps and gets stuck on a lab receive. At the same time, the labSend to terminate the sequence does not get sent by the master-lab. Thus, the slave-labs continue to run while the master-lab waits for a signal to be receive. Also while this is happening, it is possible that the master-lab actually falls asleep and fails to register any further labSends. To try and overcome this, labReceive will not be entered by the master-lab at any stage unless a message is probed first. Also a "waiting too long" time limit has been imposed such that the slave-labs will simply be abandoned if they do not quit in a timely manner.

Changed some variable names in MCMC as well.

Also updated some documentation in MCMCOptimset.