There is still one bug left where somehow the dimensions of runVar.pt is incorrect and hence doesn't fit into pt_n. This is either because either something is going wrong with runVar.pt, or with the pt_n counter itself.
To counter the later, a few lines of code have been added to dynamically increase the size of ptLocal, ptUniqLocal and logPLocal if the existing initialised matrix is completely filled up. The debug save workspace command has also been moved around to try and capture the workspace before this happens.