Overhauled parseModelm. The mechanics of generating the parameter matrix has been completely changed. Instead of requiring the user to explicitly define its structure, it is not determined implicitly based on the user generated structure of the kappa matrix. Also changed are: - Boundary and parameter description method - Parameter description generation method
The process now begins by creating a mirrored reaction structure, the param rxn structure for the reaction in question. In the mirrored form, the parameter index required is inserted as an imaginary number when a reaction parameter is free. It is negative when fixed and empty when unused. What this mirror structure does, is when the mirrorred structure is parsed, it generates a matrix that has imaginary values where the parameter is free and real otherwise. This enables a very simple way to both generate the matrix structure AND find the indices that are free parameters. This removes the need of the parameter builder in the reaction rules.

Insertion of reaction description and boundary into pFit has also been changed with the above change. A third mirrorred reaction structure, a desc structure is made and output by rxnRules. This one places descriptions into the relevant parameter field in the reaction structure while it is being classed. Doing it this way, when a free parameter is identified in the param rxn structure, it takes the corresponding value in the desc structure which gives two pieces of information encoded within a cell, where index:
1 contains which field in the boundary structure contains the default boundary
2 contains the reaction description
These both have to be defined in rxnRules. The reaction description is described below. The boundary field is defined manually by the rule set.

Reaction description has also been changed. This has now been deferred to the makeDesc function. makeDesc takes the substrate, product and enzyme list as well as a preferred parameter label as inputs. It then generates a parameter description based on a standard form.

Note the param rxn structure and the desc structure don't need to be manually generated by the user. The only explicit action that the user needs to take in creation these is substituting in the numerical parameter values in the original rxn structure (which is passed into rxnRules) and replacing them with the relevant default boundary field and reaction description and then outputing the whole structure in rxnRules