Major change in the way parseModel and parseModelm is handled
The general philosophy of this commit is parseModelm is now only responsible for creating the primitive model structure (i.e. reaction list, compartment list and species list). This primitive structure is then fed back into parseModel. This means in future when parseModelSBML is written, it will only have to create the primitive structure (the primitive structure is basically what a sigMat.m file creates). More details to follow.

parseModelm.m
  Much has been striped away from parseModelm. All of the function except for the conversion of a pre-parsed model back into it's unparsed form remains. Basically parseModelm now handles .m files as well as preparsed structures.

parseModel.m
  This now handles merging of primitive model structures, duplicate detection and resolution, and conversion of primitive model structure into matrix model structure. Most of the code used to be in parseModelm and have now just been moved here. This means the code is accessible to the output of parseModelSBML whenever that is created.

An additional flag has also been created to define whether a single structure that is parsed through needs to be reparsed or not.

Minor change in findTC to incorporate the "no reparse" flag.